Fix for STDB-11
The fix is unclear, originally the put method of CookieHandler was not called automatically as it should, and the linemode download manager had a workaround for this by callin gputManual. Starting with responses tomcat7 it appears that the cookies no longer make it to the workaround, but are passed to the put method, so we have reverted to using the put method. It is not clear what version of Java the put calling was fixed, so it is unclear if this fix will work with old Java releases, but hopefully we do not care.