Changes to support HSQLDB.
HSQLDB returns unions with literals as a fixed-size string with spaces as padding, so I changed a few UNION statements so they all use single characters when getting things back.
HSQLDB also requires the string array argument for generated columns to be in upper case (could be a bug, it was filed). This has no effect on Oracle, so all columns to be returned are now capitalized.